Moved readContributionsInput, readGraphInput,readInstanceConfig to LocalInstance
